Remove: Unused integration test
